requirements.txt
python-telegram-bot>=20.7
redis>=5.0.0
groq>=0.5.0
mistralai>=0.1.0
google-generativeai>=0.3.0
yt-dlp>=2023.12.30
beautifulsoup4>=4.12.0
requests>=2.31.0
Pillow>=10.0.0
qrcode>=7.4.0
pyzbar>=0.1.9
APScheduler>=3.10.0
aiohttp>=3.9.0
feedparser>=6.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
opencv-python>=4.8.0
